Refactoring changes
