+++ Cargo.toml
@@ -11,4 +11,3 @@ serde_derive = "1.0.79"
 tokio = { version = "0.1.0", features = ["async-await-preview"] }
 tokio-process = "0.2.2"
 futures = "0.1.23"
-structopt = "0.2"
